Feature #29: update to allow execution of DELETE and UPDATE statements
- add a Create_Statement with a query string to obtain a Delete_Statement
  or an Update_Statement
- update database drivers to allow and handle execution of direct SQL
  statements

+++ mysql/src/ado-statements-mysql.adb
@@ -1,6 +1,6 @@
 -----------------------------------------------------------------------
 --  ado-statements-mysql -- MySQL Statements
---  Copyright (C) 2009 - 2024 Stephane Carrez
+--  Copyright (C) 2009 - 2025 Stephane Carrez
 --  Written by Stephane Carrez (Stephane.Carrez@gmail.com)
 --  SPDX-License-Identifier: Apache-2.0
 -----------------------------------------------------------------------
@@ -210,8 +210,10 @@ package body ADO.Statements.Mysql is
       if Stmt.Connection = null then
          raise ADO.Sessions.Session_Error with "Database connection is closed";
       end if;
-      ADO.SQL.Append (Target => Stmt.Query.SQL, SQL => "DELETE FROM ");
-      ADO.SQL.Append_Name (Target => Stmt.Query.SQL, Name => Stmt.Table.Table.all);
+      if Stmt.Table /= null then
+         ADO.SQL.Append (Target => Stmt.Query.SQL, SQL => "DELETE FROM ");
+         ADO.SQL.Append_Name (Target => Stmt.Query.SQL, Name => Stmt.Table.Table.all);
+      end if;
       if Stmt.Query.Has_Join then
          ADO.SQL.Append (Target => Stmt.Query.SQL, SQL => Stmt.Query.Get_Join);
       end if;
@@ -271,10 +273,12 @@ package body ADO.Statements.Mysql is
       if Stmt.Connection = null then
          raise ADO.Sessions.Session_Error with "Database connection is closed";
       end if;
-      ADO.SQL.Append (Target => Stmt.This_Query.SQL, SQL => "UPDATE ");
-      ADO.SQL.Append_Name (Target => Stmt.This_Query.SQL, Name => Stmt.Table.Table.all);
-      ADO.SQL.Append (Target => Stmt.This_Query.SQL, SQL => " AS o SET ");
-      ADO.SQL.Append_Fields (Update => Stmt.This_Query);
+      if Stmt.Table /= null then
+         ADO.SQL.Append (Target => Stmt.This_Query.SQL, SQL => "UPDATE ");
+         ADO.SQL.Append_Name (Target => Stmt.This_Query.SQL, Name => Stmt.Table.Table.all);
+         ADO.SQL.Append (Target => Stmt.This_Query.SQL, SQL => " AS o SET ");
+         ADO.SQL.Append_Fields (Update => Stmt.This_Query);
+      end if;
       if Stmt.This_Query.Has_Join then
          ADO.SQL.Append (Target => Stmt.This_Query.SQL, SQL => Stmt.This_Query.Get_Join);
       end if;
